split ShapeComputationIRAnalysis
