+++ paddle/pir/dialect/shape/utils/shape_optimization_utils.h
@@ -50,6 +50,7 @@ struct SymProductHasher {
     return hash;
   }
 };
+
 class SymbolicDimMgr {
  public:
   explicit SymbolicDimMgr(ModuleOp m);
